+++ lib/rails_cursor_pagination/paginator.rb
@@ -393,7 +393,8 @@ module RailsCursorPagination
     #
     # @return [ActiveRecord::Relation]
     def relation_with_cursor_fields
-      return @relation if @relation.select_values.blank?
+      return @relation if @relation.select_values.blank? ||
+                          @relation.select_values.include?('*')
 
       relation = @relation