+++ CHANGELOG.md
@@ -15,6 +15,8 @@ These are the latest changes on the project's `master` branch that have not yet 
 --->
 
 - Adds a `limit` param to paginator that can be used instead either `first` or `last`.
+- Adds a `max_page_size` to the configuration, allowing to set a global limit to the page size (non overridable): Default `nil`
+- Supports explicitly requesting all columns without re-including the requested column
 - Require multi-factor-authentication to publish the gem on Rubygems
 
 ## [0.2.0] - 2021-04-19